Added service layer in Server:  - DataProvider uses services now (but services not implemented yet)  - added service layer data models, not implemented yet  - fixed some mistakes in wsdl files  - ids changed from int type to long type in Server and Commons

+++ Client/src/main/java/cz/cuni/mff/ufal/textan/core/Entity.java
@@ -34,7 +34,7 @@ public class Entity {
         value = entity.getValue();
         position = entity.getPosition();
         length = entity.getLength();
-        type = entity.getType();
+        type = (int) entity.getType();
     }
 
     /**
